# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcb9895f1ef4518a5382d44
  title: Todo
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /tag/:
    post:
      x-skaffolder-id: 5dcb9c44f1ef4518a5382e55
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
    get:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e41
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Tag
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
  /tag/{id}:
    delete:
      x-skaffolder-id: 5dcb9c44f1ef4518a5382e5a
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcb9c44f1ef4518a5382e52
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    post:
      x-skaffolder-id: 5dcb9c44f1ef4518a5382e54
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-resource: Tag
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Tag
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tag"
    
  #-----  Todo  -----
  /todo/:
    post:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e3d
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Todo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                details:
                  type: string
                status:
                  type: integer
                title:
                  type: string
    get:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e44
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Todo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
  /todo/{id}:
    delete:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e43
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Todo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e3e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Todo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    post:
      x-skaffolder-id: 5dcb9c3cf1ef4518a5382e3b
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Todo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Todo"
  /todo/findBycategory/{key}:
    get:
      x-skaffolder-id: 5dcb9c41f1ef4518a5382e4c
      x-skaffolder-name: findBycategory
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: findBycategory
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycategory
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycategory/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycategory
      tags:
        - Todo
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa category da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
  /todo/findBytags/{key}:
    get:
      x-skaffolder-id: 5dcb9c44f1ef4518a5382e57
      x-skaffolder-name: findBytags
      x-skaffolder-id-resource: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-resource: Todo
      x-skaffolder-crudAction: findBytags
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBytags
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBytags/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBytags
      tags:
        - Todo
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa tags da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Tag:
      x-skaffolder-id: 5dcb9b95f1ef4518a5382e30
      x-skaffolder-id-db: 5dcb9895f1ef4518a5382d45
      x-skaffolder-id-entity: 5dcb9b95f1ef4518a5382e2e
      x-skaffolder-url: /tag
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5dcb9b95f1ef4518a5382e2f
          x-skaffolder-type: String
          x-skaffolder-unique: 
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - name
    Todo:
      x-skaffolder-id: 5dcb9a4ef1ef4518a5382d96
      x-skaffolder-id-db: 5dcb9895f1ef4518a5382d45
      x-skaffolder-id-entity: 5dcb9a4ef1ef4518a5382d93
      x-skaffolder-url: /todo
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        details:
          type: string
          x-skaffolder-id-attr: 5dcb9a4ef1ef4518a5382d95
          x-skaffolder-type: String
          x-skaffolder-required: true
        status:
          type: integer
          x-skaffolder-id-attr: 5dcb9c25f1ef4518a5382e33
          x-skaffolder-type: Integer
          x-skaffolder-required: true
          x-skaffolder-enumeration:
            - DONE
            - PROGRESS
            - READY
            - REVIEW
        title:
          type: string
          x-skaffolder-id-attr: 5dcb9a4ef1ef4518a5382d94
          x-skaffolder-type: String
      x-skaffolder-relations: 
        category:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dcb9a4ef1ef4518a5382d93
          x-skaffolder-ent2: 5dcb9b07f1ef4518a5382dd5
        tags:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dcb9a4ef1ef4518a5382d93
          x-skaffolder-ent2: 5dcb9b95f1ef4518a5382e2e
      required:
        - _id
        - details
        - status
...